Add decorators for jive::GuiItem
This is a large commit because it essentially refactors the entire GuiItem class.

By splitting the different behaviours out into separate decorator classes, we allow for a more future-proof architecture which will more easily allow for additional behaviour for GuiItems. For example, it doesn't make sense to have a 'flex-wrap' property on an item with a grid layout, so we decorate a basic GuiItem in a GuiFlexContainer one, and later when we've added grid layouts we can have a GuiGridContainer decorator too.